+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTPAB34T3JK5F07FZKPNGBEF"
+        "01KTPRQR25DJ44DW9N7EF65J15"
     )
     r"""The unique identifier of the data source"""