+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTPAB34W1QHPRHKDASJ1MZHH"
+        "01KTPRQR28ZTQ8CYV7S2E9QRGC"
     )
     r"""The unique identifier of the data source"""