+++ packages/orq-rc/src/orq_ai_sdk/responses.py
@@ -278,7 +278,6 @@ class Responses(BaseSDK):
                 ),
                 sentinel="[DONE]",
                 client_ref=self,
-                data_required=False,
             )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = utils.stream_to_text(http_res)
@@ -552,7 +551,6 @@ class Responses(BaseSDK):
                 ),
                 sentinel="[DONE]",
                 client_ref=self,
-                data_required=False,
             )
         if utils.match_response(http_res, "4XX", "*"):
             http_res_text = await utils.stream_to_text_async(http_res)