+++ packages/orq-rc/RELEASES.md
@@ -8559,3 +8559,13 @@ Based on:
 - [python v4.11.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.5 - packages/orq-rc
+
+## 2026-06-09 18:03:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.773.1 (2.897.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.6 - packages/orq-rc